Move `SecretsTemplates` to a separate package
