+++ pkg/utils/sectmpl/sectmpl_test.go
@@ -1,0 +1,1 @@
+package sectmpl_test
